# Conflicts:
#	app/src/main/java/se/kth/projectarbor/project_arbor/MainService.java

+++ app/src/main/java/se/kth/projectarbor/project_arbor/Pedometer.java
@@ -166,7 +166,7 @@ class Pedometer {
     }
 
     void setHeight(float height){
-        this.height = height;
+        this.height = (double) height;
     }
 
     public void reset() {
